Done required changes on PR #2471 and added compilation after uploading .po files
