+++ app/helpers/babel.py
@@ -1,0 +1,3 @@
+from flask.ext.babel import Babel
+
+babel = Babel()          # Babel Initialization
